requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-allauth>=0.60.0
dj-rest-auth>=3.0.0
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-cors-headers>=4.3.1
